Fixed unit tests

+++ UnitTests/GW2PAO.API.UnitTest/DungeonsTableTests.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.IO;
 using GW2PAO.API.Data;
-using GW2PAO.API.Services;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace GW2PAO.API.UnitTest
@@ -9,6 +8,26 @@ namespace GW2PAO.API.UnitTest
     [TestClass]
     public class DungeonsTableTests
     {
+        private static readonly string renamedFilename = "renamedFile.xml";
+
+        [ClassInitialize]
+        public static void Init(TestContext context)
+        {
+            if (File.Exists(DungeonsTable.FileName))
+                File.Delete(DungeonsTable.FileName);
+            if (File.Exists(renamedFilename))
+                File.Delete(renamedFilename);
+
+            File.Copy(Path.Combine("TestResources", DungeonsTable.FileName), DungeonsTable.FileName);
+        }
+
+        [ClassCleanup]
+        public static void Cleanup()
+        {
+            File.Delete(DungeonsTable.FileName);
+            File.Delete(renamedFilename);
+        }
+
         [TestMethod]
         public void DungeonsTable_Constructor()
         {
@@ -23,18 +42,17 @@ namespace GW2PAO.API.UnitTest
             DungeonsTable dt = DungeonsTable.LoadTable();
             Assert.IsNotNull(dt);
             Assert.IsNotNull(dt.Dungeons);
+            Assert.IsTrue(dt.Dungeons.Count > 0);
         }
 
         [TestMethod]
         [ExpectedException(typeof(FileNotFoundException))]
         public void DungeonsTable_LoadTable_MissingFile()
         {
-            string renamedFilename = "renamedFile.xml";
             File.Move(DungeonsTable.FileName, renamedFilename);
-
             try
             {
-                DungeonsTable dt = DungeonsTable.LoadTable();
+                DungeonsTable.LoadTable();
             }
             finally
             {
@@ -46,13 +64,12 @@ namespace GW2PAO.API.UnitTest
         [ExpectedException(typeof(InvalidOperationException))]
         public void DungeonsTable_LoadTable_InvalidFile()
         {
-            string renamedFilename = "renamedFile.xml";
             File.Move(DungeonsTable.FileName, renamedFilename);
             File.WriteAllText(DungeonsTable.FileName, "invalid data");
 
             try
             {
-                DungeonsTable dt = DungeonsTable.LoadTable();
+                DungeonsTable.LoadTable();
             }
             finally
             {
@@ -64,9 +81,18 @@ namespace GW2PAO.API.UnitTest
         [TestMethod]
         public void DungeonsTable_CreateTable_Success()
         {
-            File.Delete(DungeonsTable.FileName);
-            DungeonsTable.CreateTable();
-            Assert.IsTrue(File.Exists(DungeonsTable.FileName));
+            File.Move(DungeonsTable.FileName, renamedFilename);
+            try
+            {
+                DungeonsTable.CreateTable();
+                Assert.IsTrue(File.Exists(DungeonsTable.FileName));
+            }
+            finally
+            {
+                if (File.Exists(DungeonsTable.FileName))
+                    File.Delete(DungeonsTable.FileName);
+                File.Move(renamedFilename, DungeonsTable.FileName);
+            }
         }
     }
 }
